refactor(codegen): extract std::str::* into stdlib/str.rs (Round 1.16)
21 fns lifted into StrStdlib<'ctx>: parse_{int,float,decimal}_fallible,
byte_at_unchecked, range_parse_{int,decimal}_fallible, range_eq,
index_of, can_parse_{int,float}, case_fold (lower/upper dispatcher),
substring, repeat, pad (left/right dispatcher), replace,
builder_{new,append,len,finish}, clone, from_bytes.

Surface bumps to pub(crate):
- Cx method: complete_parse_fallible_call

codegen.rs: 39914 -> 38468 (-1446 lines).
