+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -22,5 +22,6 @@ pub(crate) mod math;
 pub(crate) mod process;
 pub(crate) mod rand;
 pub(crate) mod sockopt;
+pub(crate) mod str;
 pub(crate) mod text;
 pub(crate) mod time;
